Phase 2.2: Break base_classes -> event_handler -> code_runner cycle
- Moved Event class from event_handler to gremlin/domain
- Event is now a core domain type without circular dependencies
- Created event_display_name() helper function in event_handler for display logic
- base_classes and event_handler now both import Event from domain
- Reduced circular dependencies from 5 to 3
